Rename alfalfa-client to pacer-client
Rename the library from alfalfa-client to pacer-client:
- package dir alfalfa_client/ -> pacer_client/ and module -> pacer_client.py
- class AlfalfaClient -> PacerClient
- PyPI name and references (pyproject, README, pypi_release workflow)
- docs and tests updated to the new import paths / class name

The GitHub source URLs are left unchanged. The Alfalfa* exception
classes (AlfalfaException, AlfalfaAPIException, AlfalfaWorkerException,
AlfalfaClientException) are kept as-is since they refer to the Alfalfa
server, not the client package name.
