+++ docs/index.rst
@@ -1,17 +1,17 @@
-.. Alfalfa Client documentation master file, created by
+.. Pacer Client documentation master file, created by
    sphinx-quickstart on Fri Apr  7 15:52:21 2023.
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-Alfalfa API Client Documentation
+Pacer Client Documentation
 ==========================================
-The Alfalfa API Client provides a python interface for interacting with an `Alfalfa <https://github.com/nrel/alfalfa>`_ instance.
+The Pacer Client provides a python interface for interacting with an `Alfalfa <https://github.com/nrel/alfalfa>`_ instance.
 
 .. toctree::
    :maxdepth: 2
    :caption: Contents:
 
-.. automodule:: alfalfa_client.alfalfa_client
+.. automodule:: pacer_client.pacer_client
    :members:
 
 Indices and tables
